requirements.txt
fastapi>=0.109.0
uvicorn>=0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
openai>=1.0.0
pydantic>=2.5.0,<3.0.0
pydantic[email]>=2.5.0,<3.0.0
sqlalchemy>=2.0.20
httpx>=0.25.0
supabase>=1.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0 